+++ src/main/cpp/monero_wallet_wasm_bridge.h
@@ -47,12 +47,12 @@ namespace monero_wallet_wasm_bridge
   //set<monero_wallet_listener*> get_listeners(int handle);
   void sync(int handle, emscripten::val callback);
 
-  unsigned long long get_balance_wallet(int handle);
-  unsigned long long get_balance_account(int handle, const uint32_t account_idx);
-  unsigned long long get_balance_subaddress(int handle, const uint32_t account_idx, const uint32_t subaddress_idx);
-  unsigned long long get_unlocked_balance_wallet(int handle);
-  unsigned long long get_unlocked_balance_account(int handle, const uint32_t account_idx);
-  unsigned long long get_unlocked_balance_subaddress(int handle, const uint32_t account_idx, const uint32_t subaddress_idx);
+  unsigned long get_balance_wallet(int handle);
+  unsigned long get_balance_account(int handle, const uint32_t account_idx);
+  unsigned long get_balance_subaddress(int handle, const uint32_t account_idx, const uint32_t subaddress_idx);
+  unsigned long get_unlocked_balance_wallet(int handle);
+  unsigned long get_unlocked_balance_account(int handle, const uint32_t account_idx);
+  unsigned long get_unlocked_balance_subaddress(int handle, const uint32_t account_idx, const uint32_t subaddress_idx);
   string get_accounts(int handle, bool include_subaddresses = false, const string& tag = "");
   string get_account(int handle, uint32_t account_idx, bool include_subaddresses = false);
   string get_subaddresses(int handle, const string& args);
